@@ -1,11 +1,12 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using Microsoft.Xna.Framework;
+﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using MonoGame.Extended;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace DwarvenFortification
-{	public class GridCell : ISelectableGameObject
+{
+	public class GridCell : ISelectableGameObject
 	{
 		public GridCell(CellType type)
 			=> CellType = type;
@@ -13,13 +14,16 @@ namespace DwarvenFortification
 		public void Draw(SpriteBatch sb, Point2 xy, int cellSize)
 		{
 			sb.FillRectangle(xy.X * cellSize, xy.Y * cellSize, cellSize, cellSize, CellLookup[CellType]);
-			sb.DrawString(GameServices.Fonts["Calibri"], ItemsInCell.Count.ToString(), new Vector2(xy.X * cellSize, xy.Y * cellSize), Color.Black);
+			if (ItemsInCell.Count > 0)
+			{
+				sb.DrawString(GameServices.Fonts["Calibri"], ItemsInCell.Count.ToString(), new Vector2(xy.X * cellSize, xy.Y * cellSize), Color.Black);
+			}
 		}
 
 		public override string ToString()
 			=> $"{CellType}";
 
-		public static Dictionary<CellType, Color> CellLookup = new Dictionary<CellType, Color>()
+		public static Dictionary<CellType, Color> CellLookup = new()
 		{
 			{ CellType.Grass, Color.Green },
 			{ CellType.Dirt, Color.SandyBrown },
@@ -30,7 +34,7 @@ namespace DwarvenFortification
 			{ CellType.Null, Color.Black },
 		};
 
-		public List<Item> ItemsInCell = new List<Item>();
+		public List<Item> ItemsInCell = new();
 
 		public CellType CellType
 		{
@@ -42,6 +46,11 @@ namespace DwarvenFortification
 				{
 					ItemsInCell.Clear();
 					ItemsInCell.AddRange(Enumerable.Repeat(new Item("ore"), 50));
+				}
+				if (cellType == CellType.Stone)
+				{
+					ItemsInCell.Clear();
+					ItemsInCell.AddRange(Enumerable.Repeat(new Item("stone"), 100));
 				}
 			}
 		}
