@@ -1,14 +1,17 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using DwarvenFortification.GOAP;
+﻿using DwarvenFortification.GOAP;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using MonoGame.Extended;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace DwarvenFortification
 {
 	public class Agent : ISelectableGameObject
 	{
+		static Random r = new(1);
+
 		public Agent(int x, int y, GridWorld world)
 		{
 			X = x;
@@ -16,6 +19,7 @@ namespace DwarvenFortification
 			this.world = world;
 			taskQueue = new Queue<IAgentTask>();
 			Inventory = new List<Item>();
+			Speed = ((float)r.NextDouble() * 2) + 4f;
 		}
 
 		public GridCell CurrentCell => world.CellAtXY(X, Y);
@@ -33,18 +37,27 @@ namespace DwarvenFortification
 
 		public List<GOAPAction> PossibleActions;
 
-		public int X;
-		public int Y;
+		public int X { get; set; }
+		public int Y { get; set; }
 		public Point Position
 		{
-			get => new Point(X, Y);
+			get => new(X, Y);
 			set { X = value.X; Y = value.Y; }
 		}
 
-		public float Speed = 5f;
+		public float Speed
+		{
+			get => speed * (1 - (0.25f * ((float)Inventory.Count / InventorySize)));
+			set => speed = value;
+		}
+		float speed;
 
 		public void Update(GameTime gameTime)
 		{
+			// think
+			Think();
+
+			// do
 			if (taskQueue.Count > 0)
 			{
 				var currentTask = taskQueue.Peek();
@@ -56,11 +69,51 @@ namespace DwarvenFortification
 			}
 		}
 
+		public void Think()
+		{
+			if (taskQueue.Count > 0)
+			{
+				return;
+			}
+
+			// if inventory not empty
+			if (Inventory.Count > 0)
+			{
+				var closestStorage = world.ClosestCellXYOfType(X, Y, cell => cell.CellType == CellType.Storage);
+				if (closestStorage.cell != null)
+				{
+					// move to storage
+					world.PlotPath(this, world.CoordsAtXY(closestStorage.point));
+
+					// deposit
+					AddTask(new PutDownTask(this, Inventory));
+					return;
+				}
+			}
+
+			// if can find any ores
+			var closestOre = world.ClosestCellXYOfType(X, Y, cell => cell.CellType != CellType.Storage && cell.ItemsInCell.Count > 0);
+			if (closestOre.cell != null && closestOre.cell.ItemsInCell.Count > 0)
+			{
+				// move to ore
+				world.PlotPath(this, world.CoordsAtXY(closestOre.point));
+
+				// pick up ores
+				AddTask(new PickUpTask(this, closestOre.cell.ItemsInCell.First()), InventorySize - Inventory.Count);
+				return;
+			}
+		}
+
 		public List<Item> Inventory { get; set; }
 
-		public void AddTask(IAgentTask task)
+		public int InventorySize => 5;
+
+		public void AddTask(IAgentTask task, int count = 1)
 		{
-			taskQueue.Enqueue(task);
+			for (int i = 0; i < count; ++i)
+			{
+				taskQueue.Enqueue(task);
+			}
 		}
 
 		Queue<IAgentTask> taskQueue;
@@ -74,6 +127,7 @@ namespace DwarvenFortification
 				sb.DrawRectangle(CellBounds, Color.Blue, 1);
 			}
 
+			// draw path if moving
 			var previousMoveToPoint = Position;
 			foreach (var task in taskQueue.ToArray())
 			{
@@ -82,6 +136,12 @@ namespace DwarvenFortification
 					sb.DrawLine(previousMoveToPoint.ToVector2(), mtTask.Goal.ToVector2(), Color.RosyBrown, 2);
 					previousMoveToPoint = mtTask.Goal;
 				}
+			}
+
+			// draw current task (debug)
+			if (taskQueue.TryPeek(out IAgentTask agentTask))
+			{
+
 			}
 		}
 	}
