@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace DwarvenFortification
+﻿namespace DwarvenFortification
 {
 	public interface ISelectableGameObject
 	{
